add parameter to skip remapping and contamination check

+++ modules/local/gatherstats/main.nf
@@ -9,7 +9,6 @@ process GATHERSTATS {
 
     input:
     tuple val(meta), path(general)
-    tuple val(meta2), path(contam_index)
 
     output:
     path('summary_mqc.csv'), emit: summary
@@ -21,10 +20,15 @@ process GATHERSTATS {
     def coreutils_version = '9.3'
     """
     echo '# plot_type: "generalstats"' > summary_mqc.csv
-    echo 'sample,loci,kmer_coverage,full_length,probe_length,contam_index' >> summary_mqc.csv
     cat *.stats.summary.csv | sort -k1,1 > summary-sorted.csv
-    cat *.contam_index.txt | sort -k1,1 > contam-index-sorted.csv
-    join -t ',' summary-sorted.csv contam-index-sorted.csv >> summary_mqc.csv
+    if [ -e *.contam_index.txt ]; then
+        echo 'sample,loci,kmer_coverage,full_length,probe_length,contam_index' >> summary_mqc.csv
+        cat *.contam_index.txt | sort -k1,1 > contam-index-sorted.csv
+        join -t ',' summary-sorted.csv contam-index-sorted.csv >> summary_mqc.csv
+    else
+        echo 'sample,loci,kmer_coverage,full_length,probe_length' >> summary_mqc.csv
+        cat summary-sorted.csv >> summary_mqc.csv
+    fi
 
     cat <<-END_VERSIONS > versions.yml
     "${task.process}":
@@ -34,7 +38,7 @@ process GATHERSTATS {
 
     stub:
     """
-    touch summary.csv
+    touch summary_mqc.csv
     cat <<-END_VERSIONS > versions.yml
     "${task.process}":
         coreutils: $coreutils_version
